add validate pass

+++ server.py
@@ -2,8 +2,10 @@ from flask import Flask, jsonify, request
 from flask.views import MethodView
 from models import Session, AdsTable, User
 from typing import Type
-from schema import CreateAds, PatchAds
+from schema import CreateAds, PatchAds, CreateUser, PatchUser
 from pydantic import ValidationError
+from hashlib import md5
+from sqlalchemy.exc import IntegrityError
 
 app = Flask('app')
 
@@ -30,12 +32,41 @@ def get_user(user_id: int, session: Session):
         raise HttpError(404, message='user is not found')
     return user
 
+def get_user_name(user_name, session: Session):
+    user = session.query(User).filter(User.email == user_name).all()
+    if not user:
+        raise HttpError(404, message='user is not found')
+    return user
+
 def validate(json_data, model_class: Type[CreateAds] | Type[PatchAds]):
     try:
         model_item = model_class(**json_data)
         return model_item.dict(exclude_none=True)
     except ValidationError as err:
         raise HttpError(400, err.errors())
+
+def validate_user(json_data, model_class: Type[CreateUser] | Type[PatchUser]):
+    try:
+        model_item = model_class(**json_data)
+        return model_item.dict(exclude_none=True)
+    except ValidationError as err:
+        raise HttpError(400, err.errors())
+
+def get_permission(json_data):
+    """
+    Функция проверяет соответствие поступаемых данных пользователя (email и password) с данными в базе
+    :param json_data:
+    :return: json_data
+    """
+    with Session() as session:
+        user = get_user_name(json_data.pop('username'), session)
+        password = json_data.pop('password')
+        password = password.encode()
+        hashed_password = md5(password).hexdigest()
+        if not (user[0].password == hashed_password):
+            raise HttpError(401, message='Wrong password')
+        json_data['username'] = user[0].id
+        return json_data
 
 class AdsView(MethodView):
 
@@ -52,9 +83,7 @@ class AdsView(MethodView):
 
     def post(self):
         json_data = validate(request.json, CreateAds)
-        with Session() as session:
-            user = get_user(request.json['username'], session)
-            return jsonify({'status': 'ok'})
+        json_data = get_permission(json_data)
         with Session() as session:
             new_ads = AdsTable(**json_data)
             session.add(new_ads)
@@ -69,8 +98,11 @@ class AdsView(MethodView):
 
     def patch(self, ads_id):
         json_data = validate(request.json, PatchAds)
+        json_data = get_permission(json_data)
         with Session() as session:
             ads = get_ads(ads_id, session)
+            if not(json_data['username'] == ads.username):
+                return jsonify('Wrong user')
             for field, value in json_data.items():
                 setattr(ads, field, value)
             session.commit()
@@ -83,13 +115,20 @@ class AdsView(MethodView):
             })
 
     def delete(self, ads_id):
+        json_data = validate(request.json, PatchAds)
+        json_data = get_permission(json_data)
         with Session() as session:
             ads = get_ads(ads_id, session)
+            if not(json_data['username'] == ads.username):
+                return jsonify('Wrong user')
             session.delete(ads)
             session.commit()
             return jsonify({'status': 'delete ok'})
 
 class UserView(MethodView):
+    """
+    Класс для CRUD с моделью User
+    """
 
     def get(self, user_id: int):
         with Session() as session:
@@ -101,11 +140,18 @@ class UserView(MethodView):
             })
 
     def post(self):
-        json_data = request.json
+        json_data = validate_user(request.json, CreateUser)
+        password = json_data['password']
+        password = password.encode()
+        hashed_password = md5(password).hexdigest()
+        json_data['password'] = hashed_password
         with Session() as session:
             new_user = User(**json_data)
             session.add(new_user)
-            session.commit()
+            try:
+                session.commit()
+            except IntegrityError as err:
+                raise HttpError(409, 'user already exists')
             return jsonify({
                 'id': new_user.id,
                 'email': new_user.email,
@@ -113,7 +159,7 @@ class UserView(MethodView):
             })
 
     def patch(self, user_id):
-        json_data = request.json
+        json_data = validate_user(request.json, PatchUser)
         with Session() as session:
             user = get_user(user_id, session)
             for field, value in json_data.items():
@@ -125,20 +171,13 @@ class UserView(MethodView):
                 'creationdate': user.creationdate
             })
 
-    def delete(self, user_id):
-        with Session() as session:
-            user = get_user(user_id, session)
-            session.delete(user)
-            session.commit()
-            return jsonify({'status': 'delete ok'})
-
 
 
 
 app.add_url_rule('/ads/', view_func=AdsView.as_view('ads_new'), methods=['POST'])
 app.add_url_rule('/ads/<int:ads_id>/', view_func=AdsView.as_view('ads'), methods=['GET', 'PATCH', 'DELETE'])
 app.add_url_rule('/user/', view_func=UserView.as_view('user_new'), methods=['POST'])
-app.add_url_rule('/user/<int:user_id>/', view_func=UserView.as_view('user'), methods=['GET', 'PATCH', 'DELETE'])
+app.add_url_rule('/user/<int:user_id>/', view_func=UserView.as_view('user'), methods=['GET', 'PATCH'])
 
 if __name__ == '__main__':
     app.run()
